Construct OIFCallback from actual function pointers
The FFI library used for C implementations expects pointers
to pointers. This commit moves the logic of extracting a pointer
to a function pointer to `dispatch_c` instead of doing it
on the level of C interfaces because then there is contradition
with, for example, `dispatch_python` that expects actual pointers
not pointers to pointers.