Warn about non-optimized build of Open Interfaces
Add GCC- (and Clang-) specific destructor to the dispatch library
to warn users about running a non-optimized build.
A non-optimized build is much slower than an optimized one,
so that they should not be used in production.

Mainly, this warning is useful for me, as I always forget to rebuild.