[python, bridge] Handle Fortran-style arrays better
When wrapping a Julia array in Python, it is not enough to set
the NPY_ARRAY_F_CONTIGUOUS flag to enforce column-major layout.
The order of the elements in the `strides` array that explains
the jumps between adjacent elements along each dimension must be
reversed, as it is in the opposite order to normal NumPy arrays
that use the row-major layout (C order).